+++ python/ray/data/_internal/execution/autoscaler/__init__.py
@@ -1,4 +1,5 @@
-from .autoscaler import Autoscaler, DefaultAutoscaler
+from .autoscaler import Autoscaler
+from .default_autoscaler import DefaultAutoscaler
 
 
 def create_autoscaler(topology, resource_manager, execution_id):
